+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.0.26-dev4
+
+### Enhancements
+
+* **Move default API URL parameter value to serverless API**
+* **Add check that access config always wrapped in Secret**
+* **Add togetherai embedder support**
+
+### Fixes
+
+**Fix Notion Ingestion** Fix the Notion source connector to work with the latest version of the Notion API (added `in_trash` properties to `Page`, `Block` and `Database`).
+
 ## 0.0.25
 
 ### Enhancements